requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
argcomplete==1.12.3
async-generator==1.10
attrs==21.4.0
black==21.12b0
blinker==1.5
boto3==1.20.23
botocore==1.23.23
certifi==2020.12.5
cffi==1.15.0
cfn-flip==1.3.0
chardet==4.0.0
click==7.1.2
cryptography==37.0.2
distlib==0.3.1
dnspython==2.1.0
durationpy==0.5
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.0.0
h11==0.13.0
hjson==3.0.2
idna==2.10
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
kappa
MarkupSafe==1.1.1
marshmallow==3.19.0
mongoengine==0.23.0
msgpack==1.0.2
mypy-extensions==0.4.3
outcome==1.1.0
packaging==22.0
parso==0.8.2
pathspec==0.9.0
pep517==0.12.0
pip-tools==6.4.0
pipenv==2020.11.15
placebo
platformdirs==2.4.0
pycparser==2.21
pymongo==3.11.3
pynvim==0.4.3
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.17.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==6.0
requests==2.25.1
s3transfer==0.5.0
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
trio==0.20.0
trio-websocket==0.9.2
troposphere==3.1.1
typing_extensions==4.0.1
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
webdriver-manager==3.5.4
Werkzeug==1.0.1
wsgi-request-logger==0.4.6
wsproto==1.1.0
WTForms==2.3.3
zappa
PyJWT==1.7.1
